_geomusic.c: add Fragment.import_bytes method
This method takes a `bytearray' object and converts it to the
internal floating point data format of the Fragment.  No
re-sampling or channel conversion can be made, so both the sample
rate and the number of channels of the input data and the
Fragment instance must match.  The sample width must be 16 bits
for now.  (Future versions may perform conversions and other
sample widths...).

Note: The Fragment is automatically resized to hold the data
      being imported, although it's more efficient when importing
      many small pieces to pre-size the buffer in the client code
      where the final length should be known.