dew_drop.py: use LogScale instead of hard-coded freqs
Instead of manually calculated frequencies, use LogScale
and note names to get them.  The original frequencies were
approximations of a harmonic scale, so this sounds slightly
more consonent.

The notes frequences for the scale over 3 octaves are also
printed for illustration purposes.  It's easy to change the
basic pitch or use a different scale without re-writing all
the `sine' calls.