OvertonesGenerator: add ot_decexp method
This method sets defines overtones following a decreasing
exponential.  The number of harmonics and the ratio applied to
the exponential are passed as arguments.  A small ratio will
create louder high harmonics.  The computation time will increase
with the number of harmonics to generate (default is 24) although
the resulting spectrum will be more detailed.

+ set default overtones for pure sine wave in __init__