_geomusic.c: allow levels as single float or tuples
Add `parse_levels' function to populate an array of linear
levels for each channel of a Fragment based on either a
single or a tuple of float levels in dB.  This is a
convenience as in many cases the levels are the same for
all channels.  It is used in `Fragment.amp' and in the
`overtones' dictionary.

+ update docstrings