_splat.c: define the sample type sample_t as double
Instead of using float for the sound sample type, define the
sample_t type and set it to double to avoid rounding errors.
This particularily helps when using fragments as modulation
signals, to obtain the same result as with a fixed floating point
value or as returned by a function.

When USE_V4SF is enabled, the sample type is defined to float
again.  This is to optimise for performance instead of accuracy.
There is currently no optimised mode for 64 bit (double) samples
and SSE instructions.

+ add _splat.sample_precision() module method to get the
number of bits in each floating point sample value.

Note: Due to rounding errors, the tests currently do not pass
with 32-bit (float) samples but only with the standard 64-bit
samples.