splat.interpol: add freqmod function
Add splat.interpol.freqmod function to return a Spline object
which can then be used to perform a frequency (phase) modulation
based on a set of (time, frequency) points.  This is because the
phase modulation needs to be the integral of the instantaneous
frequency.  It's also normalised with the initial frequency.