_splat.c: Fragment_mix: rename start with skip
The name `start` was ambiguous so it is now replaced by
`skip` which defines the same things i.e. time to skip in
the incoming fragment, so the skipped part doesn't get
mixed.  The `offset` argument remains unchanged and is the
time in the current fragment where the incoming one starts
to get mixed.