signal.c: fix py_float double-free
When the s->py_args refcount is decremented, all objects inside
the tuple also have their refcounts decremented.  The s->py_float
is one of them, so its refcount hits 0 at that point.
Decrementing it once more didn't cause a failure so far but it's
a bug (and it causes a segfault with Python 3.5).