add splat.tools.make to generate a makefile and run make
This is to generate a makefile which can then be passed on to
`make' to generate complex Splat pieces with several fragments in
parallel which may depend on each other and on any extra static
input files.

The convention is that a top-level "mixer" module will generate
the final sound file in a standard format such as WAV or Flac.
It will be mixing together all the other fragments generated and
saved in the Splat Audio Fragment .saf format.  Each python
module is invoked in its own process and is being passed the
output file name it needs to generate, so either a .saf for
fragment modules and a standard audio format for the mixer
module.  It's also possible to directly invoke each fragment
module and make it generate its own sound fragment in an
arbitrary format.

* add splat.tools.make to import a top-level mixer module and
  find the dependencies on other sound fragment modules with
  SPLAT_FRAGS and any other file with SPLAT_DEPS then generate a
  makefile with it

* also add splat.tools.make.main_mixer and .main_frag functions
  respectively to be used as standard main functions in top-level
  mixer and fragment modules

* add make command to tools/splat to generate a makefile for a
  given top-level module on the fly and run `make' with it in one
  command with make options passed with --opt, for example:

    splat make my_mixer --opt="-j8"

* add --print option to the splat `make' command to only print
  the generated make file and not run `make', for example to save
  it to a file:

    splat make my_mixer --print > Makefile