frag.c: add struct splat_channel and use it in splat_frag
Add a new struct splat_channel to contain the data specific to one
channel.  Use it in splat_fragment instead of having one common data
buffer for all the channels.

This allows channels to be allocated and managed individually and
helps separating the code that deals with channels from the fragments.