splat.data: add support to open and save mmap saf files
Add new Splat file format variant with the `format' attribute in the
header set to a new identifier for SAF_FORMAT_MMAP.  In this case, the
.saf file does not contain the raw flat audio data dumped from the
fragments but the pickled Fragment.info to tell which mmap files to
open in order to recreate the fragment.  This is used only if the
fragment is backed by non-temporary (persistent) mmap files.