frag.c: add open_paths arg with list of mmap paths
In preparation to support opening a series of existing mmap files, add
an open_paths argument to splat_frag_mmap with a list of mmap file
paths.  Also accept a list of paths for the mmap argument in
Fragment.__init__ and pass them on to reuse the saved data when
initialising the channels.

Update the documentation accordingly.