_splat.c: add use_mmap method to enable mmap in new frags
Add _splat.use_mmap() method to set an internal boolean flag to
use mmap by default when creating new Fragment objects when set
to True.  The flag is set to False when the module is initialised
to keep the same behaviour as before.  When called without any
arguments, the function returns the current state of the flag.

Also import _splat.use_mmap into the splat module.