frag.c: add resample support when using mmap and enable it
Resampling in general is done by making a new fragment with the
resampled data, then replacing the incoming fragment with the new
buffers and length attribute and deleting the original buffers.

When using mmap, this is essentially the same when using temporary
mmap files.  A new temporary file will be used in the resampled
fragment.

When using persistent mmap files, a temporary mmap buffer is also used
to perform the resampling.  Then the paths used in the incoming
fragment are relinked with the ones in the temporary buffers.  If the
link operation fails (not on same file system, or not supported by the
OS...), then the data is copied from the temporary files to the
persistent ones.  Finally, the incoming fragment is initialised again
with the new mmap buffers using the same file paths as before.