source.c: fix sine_table out-of-bound index error on NEON
Fix out-of-bound index in the sine look-up table due to rounding
errors using the sine_table_mask.  The fast_test.py does not typically
catch this error if the data after the end of the table happens to be
only zeroes as in this lucky case the error is very small.